requirements.txt
# ===== DEPENDENCIAS DE PRODUCCIÓN =====
fastapi==0.115.0
uvicorn[standard]==0.30.6
strawberry-graphql[fastapi]==0.246.1
mercadopago==2.2.3
python-dotenv==1.0.1
SQLAlchemy==2.0.34
pydantic==2.9.2
starlette==0.38.5
psycopg2-binary

# ===== DEPENDENCIAS DE TESTING =====
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==4.1.0
httpx==0.27.2
faker==20.1.0
sqlalchemy-utils==0.41.1